cleaned up reset
